wip refactor requests to separate classes
